rework server structure
